+++ service-implementation/src/main/resources/g2mailboxservice-production-px1.sh
@@ -1,0 +1,6 @@
+#! /bin/sh
+# ==================================================================
+# Set application specific system properties in this script
+# ==================================================================
+
+export LOGSTASHMACHINE=px1m-lvlogst11.liaison.prod