+++ service-implementation/src/main/resources/g2mailboxservice-production.sh
@@ -43,8 +43,7 @@ export HOST_IPS=$(ifconfig | grep -Eo 'inet (addr:)?([0-9]*\.){3}[0-9]*' | grep 
 export LOCAL_HOSTNAME=$(hostname)
 export DAEMON_USER=$(whoami)
 export LOGIN_USER=$(who am i | awk '{print $1}')
- 
-export LOGSTASHMACHINE=at4p-vplstash.liaison.prod
+
 export LOGSTASHPORT=4560
 export LOGSTASH_ROLLING_SIZE="250 MB"