+++ service-implementation/src/main/java/com/liaison/mailbox/dtdm/dao/ProcessorConfigurationDAO.java
@@ -164,11 +164,18 @@ public interface ProcessorConfigurationDAO extends GenericDAO<Processor> {
     long getProcessorCount();
     
     /**
-     * Retrieves downloader processors guid
+     * Retrieves processors guid
      *
-     * @return list of downloader processors guid
+     * @return list of processors guid
      */
-    List<String> getDownloadProcessorGuids(String clusterType);
+    List<String> getProcessorGuids(String clusterType, List<String> processorTypes);
+
+    /**
+     * Retrieves processors guid and process dc
+     *
+     * @return list of processors guid
+     */
+    Map<String, String> getProcessorDetails(String clusterType, List<String> processorTypes);
 
     /**
      * Retrieves processors by mailbox guid and processor name
@@ -292,18 +299,18 @@ public interface ProcessorConfigurationDAO extends GenericDAO<Processor> {
      * Update the Datacenter
      * 
      * @param dc
-     * @param ignoreDatacenters
-     * @param updateSize
+     * @param processedDC
+     * @param processSize
      */
     void updateDatacenter(String dc, List<String> processedDC, int processSize);
     
     /**
-     * Update the Downloader Datacenter by processor guid
+     * Update the Processor Datacenter by processor guid
      * 
      * @param dc
      * @param processorGuids
      */
-    void updateDownloaderDatacenter(String dc, List<String> processorGuids);
+    void updateProcessorDatacenter(String dc, List<String> processorGuids);
 
     /**
      * Update the process_dc to current_dc where the process_dc is null
@@ -432,12 +439,19 @@ public interface ProcessorConfigurationDAO extends GenericDAO<Processor> {
             .append(" AND PROCESS_DC =:" + EXISTING_PROCESS_DC).toString();
     
     String PROCESSOR_COUNT = new StringBuilder().append("SELECT COUNT(STATUS) FROM PROCESSOR")
-            .append(" WHERE STATUS <> 'DELETED' ").toString();
+            .append(" WHERE STATUS <> 'DELETED' AND PROCESS_DC != 'ALL' ").toString();
 
-    String GET_DOWNLOAD_PROCESSOR_GUIDS = new StringBuilder().append("SELECT PGUID FROM PROCESSOR")
+    String GET_PROCESSOR_GUIDS = new StringBuilder().append("SELECT PGUID FROM PROCESSOR")
             .append(" WHERE STATUS <> 'DELETED' ")
             .append(" AND CLUSTER_TYPE =:" + CLUSTER_TYPE)
-            .append(" AND TYPE = 'REMOTEDOWNLOADER'").toString();
+            .append(" AND TYPE in (:" + PROCESSOR_TYPE + ")")
+            .append(" AND PROCESS_DC != 'ALL'").toString();
+
+    String GET_PROCESSOR_DETAILS = new StringBuilder().append("SELECT PGUID, PROCESS_DC FROM PROCESSOR")
+            .append(" WHERE STATUS <> 'DELETED' ")
+            .append(" AND CLUSTER_TYPE =:" + CLUSTER_TYPE)
+            .append(" AND TYPE in (:" + PROCESSOR_TYPE + ")")
+            .append(" AND PROCESS_DC != 'ALL'").toString();
     
     String GET_ALL_DATACENTERS = new StringBuilder().append("SELECT DISTINCT PROCESS_DC FROM PROCESSOR")
             .append(" WHERE STATUS <> 'DELETED' ").toString();