+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@
 * GMB-1086 - Add Watchdog cleanup to lens event log
 * GMB-1101 - Processors' retryInterval-property should be supported
 * GGT-1581 - PX1 US UAT and PROD properties update
+* GMB-658 Mailbox updated with new 'tenancyKey' and invalid 'serviceInstance_pguid'.
 
 5.2.13
 ---------------